State: Move themes action types out
