+++ client/state/data-layer/wpcom/theme-filters/index.js
@@ -7,7 +7,7 @@ import i18n from 'i18n-calypso';
 /**
  * Internal dependencies
  */
-import { THEME_FILTERS_REQUEST, THEME_FILTERS_ADD } from 'state/action-types';
+import { THEME_FILTERS_REQUEST, THEME_FILTERS_ADD } from 'state/themes/action-types';
 import { dispatchRequest } from 'state/data-layer/wpcom-http/utils';
 import { http } from 'state/data-layer/wpcom-http/actions';
 import { errorNotice } from 'state/notices/actions';
