+++ client/state/themes/action-types.js
@@ -1,0 +1,48 @@
+export const ACTIVE_THEME_REQUEST = 'ACTIVE_THEME_REQUEST';
+export const ACTIVE_THEME_REQUEST_FAILURE = 'ACTIVE_THEME_REQUEST_FAILURE';
+export const ACTIVE_THEME_REQUEST_SUCCESS = 'ACTIVE_THEME_REQUEST_SUCCESS';
+export const RECOMMENDED_THEMES_FAIL = 'RECOMMENDED_THEMES_FAIL';
+export const RECOMMENDED_THEMES_FETCH = 'RECOMMENDED_THEMES_FETCH';
+export const RECOMMENDED_THEMES_SUCCESS = 'RECOMMENDED_THEMES_SUCCESS';
+export const THEMES_BOOKMARK_SET = 'THEMES_BOOKMARK_SET';
+export const THEMES_REQUEST = 'THEMES_REQUEST';
+export const THEMES_REQUEST_FAILURE = 'THEMES_REQUEST_FAILURE';
+export const THEMES_REQUEST_SUCCESS = 'THEMES_REQUEST_SUCCESS';
+export const THEME_ACCEPT_AUTO_LOADING_HOMEPAGE_WARNING =
+	'THEME_ACCEPT_AUTO_LOADING_HOMEPAGE_WARNING';
+export const THEME_ACTIVATE = 'THEME_ACTIVATE';
+export const THEME_ACTIVATE_FAILURE = 'THEME_ACTIVATE_FAILURE';
+export const THEME_ACTIVATE_SUCCESS = 'THEME_ACTIVATE_SUCCESS';
+export const THEME_BACK_PATH_SET = 'THEME_BACK_PATH_SET';
+export const THEME_CLEAR_ACTIVATED = 'THEME_CLEAR_ACTIVATED';
+export const THEME_DELETE = 'THEME_DELETE';
+export const THEME_DELETE_FAILURE = 'THEME_DELETE_FAILURE';
+export const THEME_DELETE_SUCCESS = 'THEME_DELETE_SUCCESS';
+export const THEME_FILTERS_ADD = 'THEME_FILTERS_ADD';
+export const THEME_FILTERS_REQUEST = 'THEME_FILTERS_REQUEST';
+export const THEME_HIDE_AUTO_LOADING_HOMEPAGE_WARNING = 'THEME_HIDE_AUTO_LOADING_HOMEPAGE_WARNING';
+export const THEME_INSTALL = 'THEME_INSTALL';
+export const THEME_INSTALL_FAILURE = 'THEME_INSTALL_FAILURE';
+export const THEME_INSTALL_SUCCESS = 'THEME_INSTALL_SUCCESS';
+export const THEME_PREVIEW_OPTIONS = 'THEME_PREVIEW_OPTIONS';
+export const THEME_PREVIEW_STATE = 'THEME_PREVIEW_STATE';
+export const THEME_REQUEST = 'THEME_REQUEST';
+export const THEME_REQUEST_FAILURE = 'THEME_REQUEST_FAILURE';
+export const THEME_REQUEST_SUCCESS = 'THEME_REQUEST_SUCCESS';
+export const THEME_SETUP_REQUEST = 'THEME_SETUP_REQUEST';
+export const THEME_SETUP_RESULT = 'THEME_SETUP_RESULT';
+export const THEME_SETUP_TOGGLE_DIALOG = 'THEME_SETUP_TOGGLE_DIALOG';
+export const THEME_TRANSFER_INITIATE_FAILURE = 'THEME_TRANSFER_INITIATE_FAILURE';
+export const THEME_TRANSFER_INITIATE_PROGRESS = 'THEME_TRANSFER_INITIATE_PROGRESS';
+export const THEME_TRANSFER_INITIATE_REQUEST = 'THEME_TRANSFER_INITIATE_REQUEST';
+export const THEME_TRANSFER_INITIATE_SUCCESS = 'THEME_TRANSFER_INITIATE_SUCCESS';
+export const THEME_TRANSFER_STATUS_FAILURE = 'THEME_TRANSFER_STATUS_FAILURE';
+export const THEME_TRANSFER_STATUS_RECEIVE = 'THEME_TRANSFER_STATUS_RECEIVE';
+export const THEME_SHOW_AUTO_LOADING_HOMEPAGE_WARNING = 'THEME_SHOW_AUTO_LOADING_HOMEPAGE_WARNING';
+export const THEME_UPLOAD_CLEAR = 'THEME_UPLOAD_CLEAR';
+export const THEME_UPLOAD_FAILURE = 'THEME_UPLOAD_FAILURE';
+export const THEME_UPLOAD_PROGRESS = 'THEME_UPLOAD_PROGRESS';
+export const THEME_UPLOAD_START = 'THEME_UPLOAD_START';
+export const THEME_UPLOAD_SUCCESS = 'THEME_UPLOAD_SUCCESS';
+export const THEMES_BANNER_HIDE = 'THEMES_BANNER_HIDE';
+export const THEMES_SHOWCASE_OPEN = 'THEMES_SHOWCASE_OPEN';
