+++ client/state/themes/actions/accept-auto-loading-homepage-warning.js
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies
  */
-import { THEME_ACCEPT_AUTO_LOADING_HOMEPAGE_WARNING } from 'state/action-types';
+import { THEME_ACCEPT_AUTO_LOADING_HOMEPAGE_WARNING } from 'state/themes/action-types';
 
 import 'state/themes/init';
