+++ client/state/themes/actions/hide-auto-loading-homepage-warning.js
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies
  */
-import { THEME_HIDE_AUTO_LOADING_HOMEPAGE_WARNING } from 'state/action-types';
+import { THEME_HIDE_AUTO_LOADING_HOMEPAGE_WARNING } from 'state/themes/action-types';
 
 import 'state/themes/init';
