+++ client/state/themes/actions/request-themes.js
@@ -8,7 +8,7 @@ import { map, property } from 'lodash';
  */
 import wpcom from 'lib/wp';
 import { fetchThemesList as fetchWporgThemesList } from 'lib/wporg';
-import { THEMES_REQUEST, THEMES_REQUEST_FAILURE } from 'state/action-types';
+import { THEMES_REQUEST, THEMES_REQUEST_FAILURE } from 'state/themes/action-types';
 import { recordTracksEvent } from 'state/analytics/actions';
 import { receiveThemes } from 'state/themes/actions/receive-themes';
 import { prependThemeFilterKeys } from 'state/themes/selectors';
