+++ client/state/themes/actions/set-theme-preview-options.js
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies
  */
-import { THEME_PREVIEW_OPTIONS } from 'state/action-types';
+import { THEME_PREVIEW_OPTIONS } from 'state/themes/action-types';
 
 import 'state/themes/init';
