+++ client/state/themes/themes-ui/actions.js
@@ -1,7 +1,11 @@
 /**
  * Internal dependencies
  */
-import { THEMES_BANNER_HIDE, THEMES_SHOWCASE_OPEN, THEMES_BOOKMARK_SET } from 'state/action-types';
+import {
+	THEMES_BANNER_HIDE,
+	THEMES_SHOWCASE_OPEN,
+	THEMES_BOOKMARK_SET,
+} from 'state/themes/action-types';
 
 import 'state/themes/init';
