+++ client/state/themes/themes-ui/reducer.js
@@ -7,7 +7,7 @@ import {
 	THEMES_BANNER_HIDE,
 	THEMES_SHOWCASE_OPEN,
 	THEMES_BOOKMARK_SET,
-} from 'state/action-types';
+} from 'state/themes/action-types';
 import { themesBannerVisibleSchema } from '../schema';
 import { combineReducers, withSchemaValidation } from 'state/utils';
