+++ client/state/ui/action-log/reducer.js
@@ -11,11 +11,11 @@ import {
 	EDITOR_PASTE_EVENT,
 	FIRST_VIEW_HIDE,
 	GUIDED_TOUR_UPDATE,
-	THEMES_REQUEST_SUCCESS,
 	PREVIEW_IS_SHOWING,
 	ROUTE_SET,
 	SITE_SETTINGS_RECEIVE,
 } from 'state/action-types';
+import { THEMES_REQUEST_SUCCESS } from 'state/themes/action-types';
 
 const relevantAnalyticsEvents = [ 'calypso_themeshowcase_theme_click' ];
